PrimitiveModelの廃止
