+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Player/PlayerCommonEntity.cs
@@ -31,7 +31,7 @@ namespace BlackSmith.Domain.Character.Player
             Name = name ?? throw new ArgumentNullException("Not found PlayerName. (O94YoFRG)");
         }
 
-        internal PlayerCommonReconstractCommand GetReconstractCommand() => new(ID, Name, Level);
+        public PlayerCommonReconstractCommand GetReconstractCommand() => new(ID, Name, Level);
 
         /// <summary>
         /// 内部情報を文字列として表示する
